[BUGFIX] Files::emptyDirectoryRecursively affects targets of symlinks
\TYPO3\FLOW3\Utility\Files::removeDirectoryRecursively() fails to
properly unlink directory symlinks on Windows.

Thus \TYPO3\FLOW3\Utility\Files::emptyDirectoryRecursively()
iterates through the target files of a symlinked folder and
removes them!

This change fixes this by introducing a custom unlink() function
that can remove a symlink regardless of it's type (file/directory).

Fixes: #32550
Releases: 1.0, 1.1
Original-Commit-Hash: 12eea665d94747288a82a9a0e02094ac6ab8d42b

+++ TYPO3.Flow/Classes/Utility/Files.php
@@ -95,27 +95,26 @@ class Files {
 	 * @see removeDirectoryRecursively()
 	 */
 	static public function emptyDirectoryRecursively($path) {
-		if (!is_dir($path)) throw new \TYPO3\FLOW3\Utility\Exception('"' . $path . '" is no directory.', 1169047616);
+		if (!is_dir($path)) {
+			throw new \TYPO3\FLOW3\Utility\Exception('"' . $path . '" is no directory.', 1169047616);
+		}
 
-		$directoryIterator = new \RecursiveDirectoryIterator($path);
-		$recursiveIterator = new \RecursiveIteratorIterator($directoryIterator);
-		foreach($recursiveIterator as $fileInfo) {
-			try {
-				if (self::is_link($path)) {
-					unlink($path);
-				} elseif (!$recursiveIterator->isDot() && unlink($fileInfo->getPathname()) === FALSE) {
+		if (self::is_link($path)) {
+			if (self::unlink($path) !== TRUE) {
+				throw new \TYPO3\FLOW3\Utility\Exception('Could not unlink symbolic link "' . $path . '".', 1323697654);
+			}
+		} else {
+			$directoryIterator = new \RecursiveDirectoryIterator($path);
+			foreach ($directoryIterator as $fileInfo) {
+				if ($fileInfo->isDir() && !$directoryIterator->isDot()) {
+					self::removeDirectoryRecursively($fileInfo->getPathname());
+				} elseif (self::unlink($fileInfo->getPathname()) !== TRUE) {
 					throw new \TYPO3\FLOW3\Utility\Exception('Could not unlink file "' . $fileInfo->getPathname() . '".', 1169047619);
 				}
-			} catch (\Exception $exception) {
-				throw new \TYPO3\FLOW3\Utility\Exception('Could not unlink file "' . $fileInfo->getPathname() . '".', 1301491043);
-			}
-		}
-		foreach ($directoryIterator as $fileInfo) {
-			if ($fileInfo->isDir() && !$directoryIterator->isDot()) {
-				self::removeDirectoryRecursively($fileInfo->getPathname());
 			}
 		}
 	}
+
 	/**
 	 * Deletes all files, directories and subdirectories from the specified
 	 * directory. Contrary to emptyDirectoryRecursively() this function will
@@ -127,17 +126,17 @@ class Files {
 	 */
 	static public function removeDirectoryRecursively($path) {
 		if (self::is_link($path)) {
-			try {
-				unlink($path);
-			} catch (\Exception $exception) {
+			if (self::unlink($path) !== TRUE) {
 				throw new \TYPO3\FLOW3\Utility\Exception('Could not unlink symbolic link "' . $path . '".', 1316000297);
 			}
 		} else {
 			self::emptyDirectoryRecursively($path);
 			try {
-				rmdir($path);
+				if (rmdir($path) !== TRUE) {
+					throw new \TYPO3\FLOW3\Utility\Exception('Could not remove directory "' . $path . '".', 1316000298);
+				}
 			} catch (\Exception $exception) {
-				throw new \TYPO3\FLOW3\Utility\Exception('Could not remove directory "' . $path . '".', 1316000298);
+				throw new \TYPO3\FLOW3\Utility\Exception('Could not remove directory "' . $path . '".', 1323961907);
 			}
 		}
 	}
@@ -256,7 +255,28 @@ class Files {
 		}
 		$normalizedPathAndFilename = strtolower(rtrim(self::getUnixStylePath($pathAndFilename), '/'));
 		$normalizedTargetPathAndFilename = strtolower(self::getUnixStylePath(realpath($pathAndFilename)));
+		if ($normalizedTargetPathAndFilename === '') {
+			return FALSE;
+		}
 		return $normalizedPathAndFilename !== $normalizedTargetPathAndFilename;
+	}
+
+	/**
+	 * A version of unlink() that works on Windows regardless on the symlink type (file/directory)
+	 *
+	 * @param string $pathAndFilename Path and name of the file or directory
+	 * @return boolean TRUE if file/directory was removed successfully
+	 */
+	static public function unlink($pathAndFilename) {
+		try {
+				// if not on Windows, call PHPs own unlink() function
+			if (DIRECTORY_SEPARATOR === '/' || is_file($pathAndFilename)) {
+				return @\unlink($pathAndFilename);
+			}
+			return rmdir($pathAndFilename);
+		} catch (\Exception $exception) {
+			return FALSE;
+		}
 	}
 }
 ?>
